+++ packages/ui-components/src/components/Translation/TranslatedEnumInput.jsx
@@ -20,9 +20,7 @@ export const TranslatedEnumField = ({
   const InputComponent = component;
 
   const translatedOptions = getTranslatedOptions(
-    Object.entries(enumValues)
-      .sort((a, b) => a[1].localeCompare(b[1]))
-      .map(([value, label]) => ({
+    Object.entries(enumValues).map(([value, label]) => ({
       value,
       label,
     })),
