+++ packages/ui-components/src/components/Translation/index.js
@@ -4,7 +4,7 @@ export { TranslatedSex } from './TranslatedSex';
 export { TranslatedText } from './TranslatedText';
 export { isTranslatedText, extractTranslationFromComponent } from './utils';
 export { getTranslatedOptions } from './getTranslatedOptions';
-export { TranslatedEnumField } from './TranslatedEnumIInput';
+export { TranslatedEnumField } from './TranslatedEnumInput';
 export { DebugTooltip } from './DebugTooltip';
 export { TranslatedOptionSelectField, TranslatedOption } from './TranslatedOptions';
 export * from './getCurrentLanguageCode';
